+++ webpack.config.js
@@ -117,8 +117,12 @@ const library_binding = function(name) {
 
 
 const config = [
-
   // library_prod(LIBRARY, SAVE_NAME, GLOBAL_JS_VAR_NAME)
+
+  library_prod(['leaflet-groupedlayercontrol/dist/leaflet.groupedlayercontrol.min.js',
+    'leaflet-groupedlayercontrol/dist/leaflet.groupedlayercontrol.min.css'],
+  'lfx-groupedlayercontrol'),
+  library_binding('lfx-groupedlayercontrol'),
 
   // "csv2geojson": "5.0.2",
   // "togeojson": "0.16.0",